Fixed the rendering parts for showing thumbnails.
Turns out that for rendering images, emacs doesn't like it
if you specify a 'display text property that contains a
'display text property.  So,

(let ((p1 (propertize "foo" 'display image))
      (p2 (propertize "bar" 'category category))
      (line (concat p1 p2)))
  (put-text-property start end 'display line))

does not behave as desired.

The fix is to split the string into two parts, the first
part explicitly displays the image and the second shows the
string as originally formatted.

I haven't thought of how this will be handled in the
future.  A simple future instance would be to include the
Twitter favorited icon instead of the current display.