Bug fix in `twitching-profile-get-image'.
Bug fix to `twitching-profile-get-image'.  Was throwing an
error when Emacs couldn't recognize the image file-type.
Specifically, in my case, Emacs didn't recognize the BMP
format and the function `image-type' in image.el was
throwing an error.

Optimization: use a special variable :nil to indicate when
an image has been downloaded but cannot be rendered by
emacs.  This saves a file system lookup for each user
profile image that has been downloaded but cannot be
rendered by Emacs.