Improved tweet display when profile images are available
Show the tweet alongside the profile image.  Slice the image
across two lines if necessary.  The images is sliced across
at most two lines with remaining, so if a tweet spans more
than two lines, the remaining lines are shown independently.

Some images, when split across two lines are shown
incompletely.  This appears to be an Emacs issue so it
probably will not be fixed.