twitching-favorite-tweet now works asynchronously
Much rejoicement and celebration.

The function get-twitching-tweet-bounds was modified to work
within the bounds of the passed buffer.  Earlier versions
worked because they were called in the same context of the
buffer they were passed as argument but that did not hold
when working asynchronously.  When called from the
asynchronous callback, (current-buffer) was the the HTTP
response buffer.